fix bugs of circle import
