+++ simpub/core/net_manager.py
@@ -1,6 +1,7 @@
+import concurrent
 import concurrent.futures
 import enum
-from typing import List, Dict, Optional
+from typing import List, Dict, Union, Optional
 from typing import Callable, TypedDict
 import asyncio
 from asyncio import sleep as async_sleep
@@ -10,12 +11,11 @@ import socket
 from socket import AF_INET, SOCK_DGRAM, SOL_SOCKET, SO_BROADCAST
 import struct
 import time
-import json
 import uuid
-import concurrent
+import abc
+from json import dumps, loads
 
 from .log import logger
-from .net_component import StringService, DictService, Service
 
 IPAddress = str
 TopicName = str
@@ -42,6 +42,159 @@ class HostInfo(TypedDict):
     topicList: List[TopicName]
 
 
+class NetComponent(abc.ABC):
+    def __init__(self):
+        if NetManager.manager is None:
+            raise ValueError("NetManager is not initialized")
+        self.manager: NetManager = NetManager.manager
+        self.running: bool = False
+        self.host_ip: str = self.manager.local_info["ip"]
+
+    def shutdown(self) -> None:
+        self.running = False
+        self.on_shutdown()
+
+    @abc.abstractmethod
+    def on_shutdown(self):
+        raise NotImplementedError
+
+
+class Publisher(NetComponent):
+    def __init__(self, topic: str):
+        super().__init__()
+        self.topic = topic
+        self.socket = self.manager.pub_socket
+        if topic in self.manager.local_info["topicList"]:
+            logger.warning(f"Host {topic} is already registered")
+        else:
+            self.manager.local_info["topicList"].append(topic)
+            logger.info(f'Publisher for topic "{self.topic}" is ready')
+
+    def publish_bytes(self, data: bytes) -> None:
+        msg = b''.join([f"{self.topic}:".encode(), b":", data])
+        self.manager.submit_task(self.send_bytes_async, msg)
+
+    def publish_dict(self, data: Dict) -> None:
+        self.publish_string(dumps(data))
+
+    def publish_string(self, string: str) -> None:
+        msg = f"{self.topic}:{string}"
+        self.manager.submit_task(self.send_bytes_async, msg.encode())
+
+    def on_shutdown(self) -> None:
+        self.manager.local_info["topicList"].remove(self.topic)
+
+    async def send_bytes_async(self, msg: bytes) -> None:
+        await self.socket.send(msg)
+
+
+class Streamer(Publisher):
+    def __init__(
+        self,
+        topic: str,
+        update_func: Callable[[], Optional[Union[str, bytes, Dict]]],
+        fps: int = 45,
+        start_streaming: bool = True,
+    ):
+        super().__init__(topic)
+        self.running = False
+        self.dt: float = 1 / fps
+        self.update_func = update_func
+        self.topic_byte = self.topic.encode("utf-8")
+        if start_streaming:
+            self.start_streaming()
+
+    def start_streaming(self):
+        self.manager.submit_task(self.update_loop)
+
+    def generate_byte_msg(self) -> bytes:
+        return dumps(
+            {
+                "updateData": self.update_func(),
+                "time": time.monotonic(),
+            }
+        ).encode("utf-8")
+
+    async def update_loop(self):
+        self.running = True
+        last = 0.0
+        try:
+            while self.running:
+                diff = time.monotonic() - last
+                if diff < self.dt:
+                    await async_sleep(self.dt - diff)
+                last = time.monotonic()
+                await self.socket.send(
+                    b"".join([self.topic_byte, b":", self.generate_byte_msg()])
+                )
+        except Exception as e:
+            logger.error(f"Error when streaming {self.topic}: {e}")
+
+
+class ByteStreamer(Streamer):
+    def __init__(
+        self,
+        topic: str,
+        update_func: Callable[[], bytes],
+        fps: int = 45,
+    ):
+        super().__init__(topic, update_func, fps)
+        self.update_func: Callable[[], bytes]
+
+    def generate_byte_msg(self) -> bytes:
+        return self.update_func()
+
+
+class Service(NetComponent):
+    def __init__(
+        self,
+        service_name: str,
+        callback: Callable[[str], Union[str, bytes, Dict]],
+    ) -> None:
+        super().__init__()
+        self.service_name = service_name
+        self.callback_func = callback
+        self.socket = self.manager.service_socket
+        # register service
+        self.manager.local_info["serviceList"].append(service_name)
+        self.manager.service_list[service_name] = self
+        self.on_trigger_events: List[Callable[[str], None]] = []
+        logger.info(f'"{self.service_name}" Service is ready')
+
+    async def callback(self, msg: str):
+        result = await asyncio.wait_for(
+            self.manager.loop.run_in_executor(
+                self.manager.executor, self.callback_func, msg
+            ),
+            timeout=5.0,
+        )
+        for event in self.on_trigger_events:
+            event(msg)
+        await self.send(result)
+
+    @abc.abstractmethod
+    async def send(self, data: Union[str, bytes, Dict]):
+        raise NotImplementedError
+
+    def on_shutdown(self):
+        pass
+
+
+class StringService(Service):
+    async def send(self, string: str):
+        await self.socket.send_string(string)
+
+
+class BytesService(Service):
+    async def send(self, data: bytes):
+        await self.socket.send(data)
+
+
+class DictService(Service):
+    async def send(self, data: Dict):
+        await self.socket.send_string(dumps(data))
+
+
 class NetManager:
 
     manager = None
@@ -50,7 +203,6 @@ class NetManager:
         self, host_ip: IPAddress = "127.0.0.1", host_name: str = "SimPub"
     ) -> None:
         NetManager.manager = self
-        self._initialized = False
         self.zmq_context = zmq.asyncio.Context()
         # subscriber
         self.sub_socket_dict: Dict[IPAddress, AsyncSocket] = {}
@@ -75,11 +227,23 @@ class NetManager:
         # client info
         self.clients: Dict[IPAddress, HostInfo] = {}
         # start the server in a thread pool
+        self.running = False
         self.executor = concurrent.futures.ThreadPoolExecutor(max_workers=5)
         self.server_future = self.executor.submit(self.start_event_loop)
-        # wait for the loop started
-        while not self.running:
-            time.sleep(0.01)
+        # wait for the loop
+        while not hasattr(self, "loop"):
+            time.sleep(0.05)
+
+    def create_socket(self, socket_type: int):
+        return self.zmq_context.socket(socket_type)
+
+    def start(self):
+        self.running = True
+
+    def start_event_loop(self):
+        self.loop = asyncio.new_event_loop()
+        asyncio.set_event_loop(self.loop)
+        # wait for the start signal
         # default service for client registration
         self.register_service = StringService(
             "Register", self.register_client_callback
@@ -93,22 +257,10 @@ class NetManager:
         self.clients_info_service = DictService(
             "GetClientInfo", self.get_clients_info
         )
-
-    def create_socket(self, socket_type: int):
-        return self.zmq_context.socket(socket_type)
-
-    def start(self):
-        self._initialized = True
-
-    def start_event_loop(self):
-        self.loop = asyncio.new_event_loop()
-        asyncio.set_event_loop(self.loop)
-        # wait for the start signal
-        while not self._initialized:
-            time.sleep(0.01)
         self.submit_task(self.broadcast_loop)
         self.submit_task(self.service_loop)
-        self.running = True
+        while not self.running:
+            time.sleep(0.01)
         self.loop.run_forever()
 
     def submit_task(
@@ -176,13 +328,13 @@ class NetManager:
         address = (broadcast_ip, ServerPort.DISCOVERY.value)
         while self.running:
             local_info = self.local_info  # update local info
-            msg = f"SimPub:{local_info['instance']}:{json.dumps(local_info)}"
+            msg = f"SimPub:{local_info['instance']}:{dumps(local_info)}"
             _socket.sendto(msg.encode(), address)
             await async_sleep(0.1)
         logger.info("Broadcasting has been stopped")
 
     def register_client_callback(self, msg: str) -> str:
-        client_info: HostInfo = json.loads(msg)
+        client_info: HostInfo = loads(msg)
         # NOTE: the client info may be updated so the reference cannot be used
         # NOTE: TypeDict is somehow block if the key is not in the dict
         # if client_info["ip"] not in self.clients:
