+++ simpub/core/simpub_server.py
@@ -6,7 +6,7 @@ import zmq
 
 from ..simdata import SimScene
 from .net_manager import init_net_manager, HostInfo
-from .net_component import Streamer, BytesService
+from .net_manager import Streamer, BytesService
 from .log import logger
 from .utils import send_message
